requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt --resolver=backtracking pyproject.toml
#
about-time==4.2.1
    # via alive-progress
aiohttp==3.8.4
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
alive-progress==3.1.4
    # via aistudent (pyproject.toml)
altair==5.0.1
    # via streamlit
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
beautifulsoup4==4.12.2
    # via aistudent (pyproject.toml)
black==23.1a1
    # via aistudent (pyproject.toml)
blinker==1.6.2
    # via streamlit
build==0.10.0
    # via pip-tools
cachetools==5.3.1
    # via
    #   google-auth
    #   streamlit
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   black
    #   pip-tools
    #   streamlit
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.7
    # via langchain
decorator==5.1.1
    # via validators
filelock==3.12.0
    # via
    #   huggingface-hub
    #   transformers
flake8==6.0.0
    # via aistudent (pyproject.toml)
fonttools==4.39.4
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.5.0
    # via huggingface-hub
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
google-auth==2.19.1
    # via
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via aistudent (pyproject.toml)
google-auth-oauthlib==1.0.0
    # via aistudent (pyproject.toml)
grapheme==0.6.0
    # via alive-progress
greenlet==2.0.2
    # via sqlalchemy
httplib2==0.22.0
    # via google-auth-httplib2
huggingface-hub==0.15.1
    # via transformers
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.6.0
    # via streamlit
isort==5.12.0
    # via aistudent (pyproject.toml)
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.3
    # via altair
kiwisolver==1.4.4
    # via matplotlib
langchain==0.0.194
    # via aistudent (pyproject.toml)
langchainplus-sdk==0.0.7
    # via langchain
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.3
    # via jinja2
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
matplotlib==3.7.1
    # via aistudent (pyproject.toml)
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
numexpr==2.8.4
    # via langchain
numpy==1.24.3
    # via
    #   altair
    #   contourpy
    #   langchain
    #   matplotlib
    #   numexpr
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
    #   transformers
oauthlib==3.2.2
    # via requests-oauthlib
openai==0.27.8
    # via aistudent (pyproject.toml)
openapi-schema-pydantic==1.2.4
    # via langchain
packaging==23.1
    # via
    #   build
    #   huggingface-hub
    #   marshmallow
    #   matplotlib
    #   streamlit
    #   transformers
pandas==2.0.2
    # via
    #   altair
    #   streamlit
pathspec==0.11.1
    # via black
pillow==9.5.0
    # via
    #   matplotlib
    #   streamlit
pip-tools==6.13.0
    # via aistudent (pyproject.toml)
platformdirs==3.5.1
    # via black
protobuf==4.23.2
    # via streamlit
pyarrow==12.0.0
    # via streamlit
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycodestyle==2.10.0
    # via flake8
pydantic==1.10.9
    # via
    #   langchain
    #   langchainplus-sdk
    #   openapi-schema-pydantic
pydeck==0.8.1b0
    # via streamlit
pyflakes==3.0.1
    # via flake8
pygments==2.15.1
    # via rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via
    #   httplib2
    #   matplotlib
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   streamlit
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   aistudent (pyproject.toml)
    #   huggingface-hub
    #   langchain
    #   transformers
regex==2023.6.3
    # via
    #   tiktoken
    #   transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   langchain
    #   langchainplus-sdk
    #   openai
    #   requests-oauthlib
    #   streamlit
    #   tiktoken
    #   transformers
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rich==13.4.1
    # via streamlit
rsa==4.9
    # via google-auth
safetensors==0.3.1
    # via transformers
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
smmap==5.0.0
    # via gitdb
soupsieve==2.4.1
    # via beautifulsoup4
sqlalchemy==2.0.15
    # via langchain
streamlit==1.23.1
    # via
    #   aistudent (pyproject.toml)
    #   streamlit-elements
    #   streamlit-lottie
streamlit-elements==0.1.0
    # via aistudent (pyproject.toml)
streamlit-lottie==0.0.5
    # via aistudent (pyproject.toml)
tenacity==8.2.2
    # via
    #   aistudent (pyproject.toml)
    #   langchain
    #   langchainplus-sdk
    #   streamlit
tiktoken==0.4.0
    # via aistudent (pyproject.toml)
tokenizers==0.13.3
    # via transformers
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.2
    # via streamlit
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.30.0
    # via aistudent (pyproject.toml)
typing-extensions==4.6.3
    # via
    #   huggingface-hub
    #   pydantic
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3
    # via streamlit
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
validators==0.20.0
    # via streamlit
watchdog==3.0.0
    # via aistudent (pyproject.toml)
wheel==0.40.0
    # via pip-tools
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
